scrape/requirements.txt
pydantic==1.10.7
pydantic[email]
redis==4.5.4
asyncssh==2.13.1
aiosmtplib==2.0.1
APScheduler==3.10.1
pandas==2.0.0
matplotlib==3.7.1
premailer==3.10.0
croniter==1.3.14
tortoise-orm==0.19.3
aiomysql==0.1.1